play some dorian mode scales
