+++ part2/experiment02.py
@@ -2,15 +2,25 @@
 # SPDX-License-Identifier: MIT
 
 # copy to the CIRCUITPY drive as "code.py"
-# also copy midi_player.py there
-"""Part 2 Experiment 01 -- bring back the light show"""
+"""Part 2 Experiment 02 -- other modes"""
 import time
 import board
 
 from neopixel import NeoPixel
 from midi_player import MidiPlayer
 
+# thanks https://github.com/adafruit/Adafruit_Learning_System_Guides
+# /blob/f44f5f961ceb9dccb785c5843c22b12bcc1f7b5b/MIDI_Modal_Keyboard/code.py
+major = (0, 2, 4, 5, 7, 9, 11)
+minor = (0, 2, 3, 5, 7, 8, 10)
+dorian = (0, 2, 3, 5, 7, 9, 10)
+phrygian = (0, 1, 3, 5, 7, 8, 10)
+lydian = (0, 2, 4, 6, 7, 9, 11)
+mixolydian = (0, 2, 4, 5, 7, 9, 10)
+locrian = (0, 1, 3, 5, 6, 8, 10)
+
 player = MidiPlayer()
+player.riff = [60 + m for m in dorian]
 pixel = NeoPixel(board.NEOPIXEL, 1, brightness=0.3)
 lightshow = [0xFF0000, 0xCC0000, 0x990000, 0x660000, 0x330000, 0x000000]
 while True:
